requirements.txt
opensmile~=2.4.1
pandas~=1.1.5
numpy~=1.21.6
praat-parselmouth~=0.4.3